Change `(∧)` to better match `pkgs.lib.recursiveUpdate`.  Fixes #1
This changes the generated code for `(∧)` to prefer the right attribute
if both arguments share the same key but the correspond values are not
both attribute sets.

Note that in theory this is not even possible in Dhall.  The `(∧)`
operator does not permit you to combine records if they have conflicting
fields that are not both records.  However, in theory you could compile
Dhall code like this to Nix:

```
λ(x : {}) → λ(y : {}) → x ∧ y
```

... and then use the generated Nix code to combine two records with
conflicting fields.  Nix would permit this since the type annotation on
the function arguments is lost in the conversion to Nix.  So even though
users shouldn't be using the code in this way we still care about
gracefully handling the case where users misuse this.

For the case where fields conflict we match the precedence of
`pkgs.lib.recursiveUpdate` which prefers the right field.  `(∧)` is
spiritually modeled after `pkgs.lib.recursiveUpdate` so it should match
the precedence.